requirements.txt
pandas==2.2.3
numpy==2.1.3
pyarrow==18.0.0
openpyxl==3.1.5
xlsxwriter==3.2.0
PyYAML==6.0.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
scikit-learn==1.5.2
statsmodels==0.14.4
plotly==5.24.1
jinja2==3.1.4
tqdm==4.67.1
streamlit==1.32.2
google-api-python-client==2.160.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
